after MReview

+++ src/main/java/mate/academy/springboot/datajpa/service/impl/CategoryServiceImpl.java
@@ -25,6 +25,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new NoSuchElementException("Can`t find category by id: " + id);
+        }
         categoryRepository.deleteById(id);
     }
